Compilation under MSVC, enable Start Center Community on Mac, Linux and Windows compiled with MSVC, support toolchain to develop MuseScore in Visual Studio Community.

+++ zerberus/channel.cpp
@@ -30,8 +30,8 @@ Channel::Channel(Zerberus* ms, int i)
       _instrument = 0;
       _gain       = 1.0;
       _midiVolume = 1.0;
-      _panLeftGain  = cosf(M_PI_2 * 64.0/126.0);
-      _panRightGain = sinf(M_PI_2 * 64.0/126.0);
+      _panLeftGain  = cosf(static_cast<float>(M_PI_2 * 64.0/126.0));
+      _panRightGain = sinf(static_cast<float>(M_PI_2 * 64.0/126.0));
       memset(ctrl, 0, 128 * sizeof(char));
       ctrl[Ms::CTRL_EXPRESSION] = 127;
       }
